+++ src/bezel/invocations.ts
@@ -44,6 +44,7 @@ import { BzlConfiguration, InvocationsConfiguration } from './configuration';
 import {
   BzlFrontendLinkItem,
   DisabledItem,
+  DocumentationLinkItem,
   Expandable,
   Revealable,
   RunnableComponentItem,
@@ -154,6 +155,7 @@ export class InvocationsItem
 
   async getChildrenInternal(): Promise<vscode.TreeItem[]> {
     const items: vscode.TreeItem[] = [];
+    items.push(new DocumentationLinkItem('invocations'))
 
     if (this.invocations.status === Status.DISABLED) {
       items.push(new DisabledItem('Depends on the Bzl Service'));